+++ project_files/app.rb
@@ -1,8 +1,8 @@
-require './book'
-require './student'
-require './teacher'
-require './rental'
-require './classroom'
+require_relative 'book'
+require_relative 'student'
+require_relative 'teacher'
+require_relative 'rental'
+require_relative 'classroom'
 
 class PersonsManager
   def initialize
