+++ project_files/console_manager.rb
@@ -1,4 +1,4 @@
-require './app'
+require_relative 'app'
 
 class ConsoleManager
   def initialize
