+++ project_files/person.rb
@@ -1,4 +1,4 @@
-require './nameable'
+require_relative 'nameable'
 
 class Person < Nameable
   attr_accessor :name, :age, :id, :has_rented
