+++ project_files/student.rb
@@ -1,5 +1,5 @@
 require 'json'
-require './person'
+require_relative 'person'
 
 class Student < Person
   attr_reader :classroom
