+++ project_files/trimmer.rb
